requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.6
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.5
idna==3.4
importlib-resources==6.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.17.0
